iniciativas-backup/simple-backup-app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.3
boto3==1.34.122
jinja2==3.1.4
pyyaml==6.0.1
pydantic==2.7.4
